+++ sdk/src/core/Cache-spec.ts
@@ -4,11 +4,11 @@ describe('RingCentral.core.Cache', () => {
 
     describe('getItem', () => {
 
-        it('returns null if item not found', asyncTest(sdk => {
+        it('returns null if item not found', asyncTest(async sdk => {
 
             const cache = sdk.cache();
 
-            expect(cache.getItem('foo')).to.equal(null);
+            expect(await cache.getItem('foo')).to.equal(null);
 
         }));
 
@@ -16,16 +16,17 @@ describe('RingCentral.core.Cache', () => {
 
     describe('setItem', () => {
 
-        it('sets an item in storage', asyncTest(sdk => {
+        it('sets an item in storage', asyncTest(async sdk => {
 
             const cache = sdk.cache();
             const json = {foo: 'bar'};
 
-            expect(cache.setItem('foo', json).getItem('foo')).to.deep.equal(json);
+            await cache.setItem('foo', json);
+            expect(await cache.getItem('foo')).to.deep.equal(json);
 
-            cache.removeItem('foo');
+            await cache.removeItem('foo');
 
-            expect(cache.getItem('foo')).to.equal(null);
+            expect(await cache.getItem('foo')).to.equal(null);
 
         }));
 
@@ -33,19 +34,19 @@ describe('RingCentral.core.Cache', () => {
 
     describe('clean', () => {
 
-        it('removes all prefixed entries from cache leaving non-prefixed ones untouched', asyncTest(sdk => {
+        it('removes all prefixed entries from cache leaving non-prefixed ones untouched', asyncTest(async sdk => {
 
             const cache = sdk.cache();
 
-            cache['_externals'].localStorage['rc-foo'] = '"foo"';
-            cache['_externals'].localStorage.foo = '"foo"';
+            cache['_externals'].localStorage.setItem('rc-foo', '"foo"');
+            cache['_externals'].localStorage.setItem('foo', '"foo"');
 
-            expect(cache.getItem('foo')).to.equal('foo');
+            expect(await cache.getItem('foo')).to.equal('foo');
 
-            cache.clean();
+            await cache.clean();
 
-            expect(cache.getItem('foo')).to.equal(null);
-            expect(cache['_externals'].localStorage.foo).to.equal('"foo"');
+            expect(await cache.getItem('foo')).to.equal(null);
+            expect(cache['_externals'].localStorage.getItem('foo')).to.equal('"foo"');
 
         }));
 
@@ -53,15 +54,15 @@ describe('RingCentral.core.Cache', () => {
 
     describe('prefix', () => {
 
-        it('different prefixes dont overlap', asyncTest(sdk1 => {
+        it('different prefixes dont overlap', asyncTest(async sdk1 => {
 
             const sdk2 = createSdk({cachePrefix: 'foo'});
 
             const cache1 = sdk1.cache();
 
-            cache1.setItem('foo', {foo: 'bar'});
+            await cache1.setItem('foo', {foo: 'bar'});
 
-            expect(sdk2.cache().getItem('foo')).to.equal(null);
+            expect(await sdk2.cache().getItem('foo')).to.equal(null);
 
         }));