+++ sdk/src/core/Externals.ts
@@ -10,13 +10,23 @@ export interface ExternalsOptions {
     localStorage?: Storage
 }
 
+let storage = null;
+
+const getStorage = () => {
+    if (typeof root.localStorage !== 'undefined') return root.localStorage;
+    if (storage) return storage;
+    const DomStorage = require('dom-storage');
+    storage = new DomStorage(null, {strict: true});
+    return storage;
+};
+
 export default class Externals implements ExternalsOptions {
 
     fetch = root.fetch;
     Request = root.Request;
     Response = root.Response;
     Headers = root.Headers;
-    localStorage = (typeof root.localStorage !== 'undefined') ? root.localStorage : {};
+    localStorage = getStorage();
 
     constructor({
                     fetch: fetchImpl,