+++ sdk/webpack.config.js
@@ -5,5 +5,14 @@ module.exports = createConfig({
     filename: 'ringcentral',
     entry: './src/SDK.ts',
     outputPath: path.resolve(__dirname, 'dist'),
-    libraryName: 'sdk'
+    libraryName: 'sdk',
+    externals: {
+        'dom-storage': {
+            commonjs: 'dom-storage',
+            commonjs2: 'dom-storage',
+            amd: null,
+            root: null
+        }
+    }
+
 });