+++ subscriptions/src/subscription/Subscription-spec.ts
@@ -321,7 +321,7 @@ describe('RingCentral.subscription.Subscription', function() {
                 }
             });
 
-            expect(subscription._decrypt(aesMessage)).to.deep.equal({
+            expect(subscription['_decrypt'](aesMessage)).to.deep.equal({
                 "timestamp": "2014-03-12T20:47:54.712+0000",
                 "body": {
                     "extensionId": 402853446008,