+++ CHANGELOG.md
@@ -6,6 +6,8 @@ This document contains only major functionality changes in certain versions. Min
 
 - SDK was separated into 2 packages: `@ringcentral/sdk` and `@ringcentral/subscriptions`
 - `appKey` and `appSecret` were replaced with `clientId` and `clientSecret`
+- Added async implementation of `Storage`
+- Platform method `createUrl('xxx', {addToken: true})` has been extracted as separate async method `signUrl('xxx')`
 
 ## 3.0